+++ src/pydrobert/torch/_estimators.py
@@ -12,14 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import abc
+
 from typing import Callable
+
 import torch
 
 
 FunctionOnSample = Callable[[torch.Tensor], torch.Tensor]
 """Type for functions of samples used in estimators
 
-This type is intended for use in estimators from :mod:`pydrobert.torch.estimators`.
+This type is intended for use in estimators subclassing :class:`Estimator`.
 
 A `FunctionOnSample` is a callable which accepts a :class:`torch.Tensor` and returns a
 :class:`torch.Tensor`. The input is of shape ``(N,) + batch_size + event_size``, where
@@ -27,4 +30,84 @@ A `FunctionOnSample` is a callable which accepts a :class:`torch.Tensor` and ret
 the proposal distribution. The return value is a tensor which broadcasts with ``(N,) +
 batch_size``, usually of that shape, storing the values of the function evaluated on
 each sample.
+
+`FunctionOnSample` can be a :class:`torch.nn.Module`.
 """
+
+
+class Estimator(metaclass=abc.ABCMeta):
+    r"""Computes an estimate of an expectation
+
+    An estimator estimates the value of a function :math:`f` integrated over a
+    probability density :math:`P`
+
+    .. math::
+
+        z = \mathbb{E}_{b \sim P}\left[f(b)\right]
+          = \int_{b \in \mathrm{supp}(P)} f(b) \mathrm{d}P(b).
+    
+    The value of :math:`z` can be estimated in many ways. This base class serves as the
+    common foundation for those estimators. The usage pattern is as follows:
+
+    .. code-block:: python
+
+        def func(b):
+            # return the value of f(b) here
+        
+        # ...
+        # training loop
+        for epoch in range(num_epochs):
+            # ...
+            # 1. Determine parameterization (e.g. logits) from inputs.
+            # 2. Initialize the distribution and estimator in the training loop.
+            dist = torch.distributions.SomeDistribution(logits=logits)
+            estimator = pydrobert.torch.estimators.SomeEstimator(dist, func, ...)
+            z = estimator()  # of shape dist.batch_shape
+            # 3. calculate loss as a function of z
+            loss.backwards()
+            # ...
+
+    Parameters
+    ----------
+    proposal : torch.distributions.Distribution
+        The distribution over which the expectation is taken. This is usually but not
+        always :math:`P` (see :class:`ImportanceSamplingEstimator` for a
+        counterexample).
+    func : FunctionOnSample
+        The function :math:`f`.
+    is_log : bool, optional
+        If :obj:`True`, `func` defines :math:`\log f` instead of :math:`f`. Unless
+        otherwise specified, `is_log` being true is semantically identical to redefining
+        `func` as::
+
+            def new_func(b):
+                return func(b).exp()
+        
+        and setting `is_log` to :obj:`False`. Practically, `is_log` may improve the
+        numerical stability of certain estimators.
+    
+    Notes
+    -----
+    An estimator is not a :class:`torch.nn.Module` and is not in general safe to be
+    JIT scripted or traced. The parameterization of the proposal distribution is usually
+    output 
+    """
+
+    proposal: torch.distributions.Distribution
+    func: FunctionOnSample
+    is_log: bool
+
+    def __init__(
+        self,
+        proposal: torch.distributions.Distribution,
+        func: FunctionOnSample,
+        is_log: bool = False,
+    ):
+        super().__init__()
+        self.proposal = proposal
+        self.func = func
+        self.is_log = is_log
+
+    @abc.abstractmethod
+    def __call__(self) -> torch.Tensor:
+        raise NotImplementedError